tests/wpt/tests/css/css-shapes/parsing/shape-outside-valid-position.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Shapes Module Level 1: parsing shape-outside with valid position values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#typedef-position">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8695">
<meta name="assert" content="shape-outside positions support the full '<position>' grammar.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("shape-outside", "circle(at 10%)", "circle(at 10% center)");
test_valid_value("shape-outside", "circle(at 20% 30px)");
test_valid_value("shape-outside", "circle(at 30px center)", "circle(at 30px center)");
test_valid_value("shape-outside", "circle(at 40px top)", "circle(at 40px top)");
test_valid_value("shape-outside", "circle(at bottom 10% right 20%)", "circle(at right 20% bottom 10%)");
test_valid_value("shape-outside", "circle(at bottom right)", "circle(at right bottom)");
test_valid_value("shape-outside", "circle(at center)", "circle(at center center)");
test_valid_value("shape-outside", "circle(at center 50px)", "circle(at center 50px)");
test_valid_value("shape-outside", "circle(at center bottom)", "circle(at center bottom)");
test_valid_value("shape-outside", "circle(at center center)", "circle(at center center)");
test_valid_value("shape-outside", "circle(at center left)", "circle(at left center)");
test_valid_value("shape-outside", "circle(at left)", "circle(at left center)");
test_valid_value("shape-outside", "circle(at left bottom)", "circle(at left bottom)");
test_valid_value("shape-outside", "circle(at left center)", "circle(at left center)");
test_valid_value("shape-outside", "circle(at right 40%)", "circle(at right 40%)");
test_valid_value("shape-outside", "circle(at right 30% top 60px)", "circle(at right 30% top 60px)");
test_valid_value("shape-outside", "circle(at top)", "circle(at center top)");
test_valid_value("shape-outside", "circle(at top center)","circle(at center top)");
test_valid_value("shape-outside", "circle(at top 0% right calc(10% * sign(1em - 1px)))", "circle(at right calc(10% * sign(1em - 1px)) top 0%)");
test_valid_value("shape-outside", "circle(at top 0% right calc(10% * sibling-index()))", "circle(at right calc(10% * sibling-index()) top 0%)");
</script>
</body>
</html>
